Update changelog
Fixing the output turned out to not be necessary in the end. The
failures I was getting when using format=nunit2 were related to
a bug in [NUnit](https://github.com/nunit/nunit-console/issues/775)
where it didn't detect and load extensions when the NUnit package
was installed in the global package location.