Fix failure when running with `dotnet pack`
Previously we were missing subfolders under `buildMultiTargeting`.
This caused .NET Core to attempt to use the net461 version which
failed because of a missing assembly. This change properly creates
`net461` and `netstandard2.0` subfolders, allowing .NET Core to use
the netstandard version.

This change also creates empty files under `lib`. This removes a
warning and allows nuget.org to properly detect the supported
frameworks.