Support Keep a Changelog format input
- SIL.ReleaseTasks expected a different format of changelog. Support
keepachangelog.com style changelogs as well, which have a preamble,
[Unreleased] version, and sub-headings for types of changes (eg
"Added", "Changed", ...).
- Remove the preamble and [Unreleased] heading, to not carry those
thru to the user.
- There is some unfortunate mixing of string, string[], and
List<string> that could use some smoothing out.
- CreateChangelogEntry.cs: Setting EntryDate to default to
DateTime.Now will work fine when the class is created right before
used, and was a way to make sure the build never fails because the
unit tests were 1 s off. But it could have unexpected results if one
were to ever use this class in a long-running system.